refactor: extract shared DetailSheet shell
Move the bottom-sheet plumbing (modal transition, backdrop, blur, panel
animation, drag-to-dismiss, scroll area, optional edit row) into a shared
component at src/shared/components/DetailSheet.tsx. EventDetailSheet
becomes a thin wrapper: content-only, passes `editLabel`+`onEdit` for the
edit row. No behavior change.

Sets up check-side usage if we ever give interest checks a tap-to-detail
UX (final step of the unification series) — but since check cards already
surface all their info inline with the shared InlineCommentsBox, that's
not strictly needed yet.
